+++ backend/api/__init__.py
@@ -1,0 +1,1 @@
+"""Expose API routes via FastAPI routers from this package."""
