+++ backend/api/user.py
@@ -1,0 +1,76 @@
+from fastapi import APIRouter, HTTPException, Depends
+from ..services import UserService, ChallengeService
+from ..models import User, Challenge
+
+api = APIRouter()
+
+
+# ----------USER API ROUTES----------------
+#api route retrieves ALL registered users
+@api.get("/api/registrations")
+def get_registrations(user_service: UserService = Depends()) -> list[User]:
+    return user_service.all()
+
+#api route registers a new user
+@api.post("/api/registrations")
+def new_registration(user: User, user_service: UserService = Depends()) -> User:
+        try:
+            return user_service.create(user)
+        except Exception as e:
+            raise HTTPException(status_code=422, detail=str(e))
+
+#api route retrieves user given email
+@api.get("/api/users/{email}", responses={404: {"model": None}})
+def get_user(email: str, user_service: UserService = Depends()) -> User:
+    try: 
+        return user_service.get(email)
+    except Exception as e:
+        raise HTTPException(status_code=404, detail=str(e))
+
+#api route to update user info
+#TODO: Change update function use PUT instead of POST
+@api.post("/api/users")
+def update_user(user: User, user_service: UserService = Depends()) -> User:
+    try:
+        return user_service.update(user)
+    except Exception as e:
+        raise HTTPException(status_code=422, detail=str(e))
+
+#api route deletes registered user
+@api.delete("/api/delete/users/{email}")
+def delete_user(email: str, user_service = Depends(UserService)) -> User:
+    try:
+        return user_service.delete(email)
+    except Exception as e:
+        raise HTTPException(status_code=404, detail=str(e))
+
+# ----------CHALLENGE API ROUTES----------------
+#api route retrieves ALL challenges
+@api.get("/api/challenges")
+def get_challenges(challenge_service: ChallengeService = Depends()) -> list[Challenge]:
+    return challenge_service.all()
+
+#api route registers a new challenge
+@api.post("/api/challenges")
+def new_challenge(challenge: Challenge, challenge_service: ChallengeService = Depends()) -> Challenge:
+        try:
+            return challenge_service.create(challenge)
+        except Exception as e:
+            raise HTTPException(status_code=422, detail=str(e))
+        
+#api route retrieves challenge given id
+#TODO: implement a way to find challenge and get the correct id
+@api.get("/api/challenges/{id}", responses={404: {"model": None}})
+def get_challenge(id: int, challenge_service: ChallengeService = Depends()) -> Challenge:
+    try: 
+        return challenge_service.get(id)
+    except Exception as e:
+        raise HTTPException(status_code=404, detail=str(e))
+    
+#api route deletes challenge FROM THE DATABASE
+@api.delete("/api/delete/challenges/{id}")
+def delete_challenge(id: int, challenge_service = Depends(ChallengeService)) -> Challenge:
+    try:
+        return challenge_service.delete(id)
+    except Exception as e:
+        raise HTTPException(status_code=404, detail=str(e))
